Add files via upload
A simple and effective saliency detection approach
Hui Zhang†, WeiqiangWang†‡, Guiping Su†, Lijuan Duan
Abstract:
This paper presents a simple and effective method to compute the pixel saliency with full resolution in an image. First, the proposed method creates an image representation of four color channels through the modified computation on the basis of Itti et al.[5]. Then the most informative channel is automatically identified from the derived four color channels. Finally, the pixel saliency is computed through the simple combination of contrast feature and spatial attention function on the individual channel. The proposed method is computationally very simple, but it achieves a very good performance in the comparison experiments with six other saliency detection methods. On the challenging database with 1,000 images, it outperforms six other methods in both identifying salient pixels and segmenting salient regions.